test(rust): split graph invariant coverage
