+++ crates/organizational-graph/src/lib.rs
@@ -424,680 +424,4 @@ impl GraphRead for OrganizationalGraph {
 }
 
 #[cfg(test)]
-mod tests {
-    use cerebro_organizational_model::{
-        AssertionProvenance, CanonicalIdentity, CanonicalIdentityId, CollectionId,
-        CompleteCollection, Entity, EntityKind, GraphAssertion, IdentityBindingAssertion,
-        IdentityBindingState, IdentityClaim, IdentityClaimKind, IdentityResolutionMethod,
-        ObservationId, ObservationRef, ProviderIdentity, ProviderKind, RelationKind,
-        RelationshipAssertion, SourceRuntimeId,
-    };
-
-    use super::*;
-
-    fn identity_delta(
-        provider_id: &str,
-        canonical_id: &str,
-        state: IdentityBindingState,
-        claim: Option<IdentityClaim>,
-    ) -> GraphDelta {
-        identity_delta_with_method(
-            provider_id,
-            canonical_id,
-            IdentityResolutionMethod::HumanDecision,
-            state,
-            claim,
-        )
-    }
-
-    fn identity_delta_with_method(
-        provider_id: &str,
-        canonical_id: &str,
-        method: IdentityResolutionMethod,
-        state: IdentityBindingState,
-        claim: Option<IdentityClaim>,
-    ) -> GraphDelta {
-        let tenant = TenantId::parse("tenant-a").unwrap();
-        let runtime = SourceRuntimeId::parse("okta-prod").unwrap();
-        let collection = CompleteCollection::new(
-            tenant.clone(),
-            runtime.clone(),
-            CollectionId::parse(format!("collection-{canonical_id}")).unwrap(),
-            "okta.users",
-            10,
-        )
-        .unwrap();
-        let provider = ProviderIdentity::new(
-            tenant.clone(),
-            runtime,
-            ProviderKind::parse("okta.user").unwrap(),
-            provider_id,
-            "Provider Person",
-        )
-        .unwrap();
-        let canonical = CanonicalIdentity::new(
-            tenant,
-            CanonicalIdentityId::parse(canonical_id).unwrap(),
-            "Canonical Person",
-        )
-        .unwrap();
-        let provenance = AssertionProvenance::direct(
-            vec![
-                ObservationRef::new(
-                    collection.receipt(),
-                    ObservationId::parse(format!("observation-{canonical_id}")).unwrap(),
-                    format!("okta.user:{provider_id}"),
-                )
-                .unwrap(),
-            ],
-            "okta-identity-mapper",
-            "v1",
-        )
-        .unwrap();
-        let binding = IdentityBindingAssertion::new(
-            &provider, &canonical, method, claim, state, provenance, 10,
-        )
-        .unwrap();
-        let mut builder = collection.begin_delta();
-        builder.add_entity(provider.into_entity()).unwrap();
-        builder.add_entity(canonical.into_entity()).unwrap();
-        builder
-            .add_assertion(GraphAssertion::IdentityBinding(binding))
-            .unwrap();
-        builder.build()
-    }
-
-    fn workforce_delta(employee_id: &str, email: &str) -> (GraphDelta, CanonicalIdentity) {
-        let tenant = TenantId::parse("tenant-a").unwrap();
-        let runtime = SourceRuntimeId::parse("okta-prod").unwrap();
-        let collection = CompleteCollection::new(
-            tenant.clone(),
-            runtime.clone(),
-            CollectionId::parse(format!("okta-{employee_id}")).unwrap(),
-            "okta.users",
-            10,
-        )
-        .unwrap();
-        let provider = ProviderIdentity::new(
-            tenant.clone(),
-            runtime,
-            ProviderKind::parse("okta.identity_user").unwrap(),
-            "00u1",
-            "Provider Person",
-        )
-        .unwrap();
-        let employee_claim = IdentityClaim::employee_id(employee_id).unwrap();
-        let canonical =
-            CanonicalIdentity::for_claim(tenant, &employee_claim, "Canonical Person").unwrap();
-        let provenance = AssertionProvenance::direct(
-            vec![
-                ObservationRef::new(
-                    collection.receipt(),
-                    ObservationId::parse(format!("observation-{employee_id}")).unwrap(),
-                    "okta.user:00u1",
-                )
-                .unwrap(),
-            ],
-            "okta-identity-mapper",
-            "v1",
-        )
-        .unwrap();
-        let employee_binding = IdentityBindingAssertion::new(
-            &provider,
-            &canonical,
-            IdentityResolutionMethod::AuthoritativeEmployeeId,
-            Some(employee_claim),
-            IdentityBindingState::Confirmed,
-            provenance.clone(),
-            10,
-        )
-        .unwrap();
-        let email_binding = IdentityBindingAssertion::new(
-            &provider,
-            &canonical,
-            IdentityResolutionMethod::VerifiedEmail,
-            Some(IdentityClaim::verified_email(email).unwrap()),
-            IdentityBindingState::Confirmed,
-            provenance,
-            10,
-        )
-        .unwrap();
-        let mut builder = collection.begin_delta();
-        builder.add_entity(provider.into_entity()).unwrap();
-        builder.add_entity(canonical.clone().into_entity()).unwrap();
-        builder
-            .add_assertion(GraphAssertion::IdentityBinding(employee_binding))
-            .unwrap();
-        builder
-            .add_assertion(GraphAssertion::IdentityBinding(email_binding))
-            .unwrap();
-        (builder.build(), canonical)
-    }
-
-    fn claim_match_delta(
-        source: &str,
-        provider_id: &str,
-        email: &str,
-        canonical: &CanonicalIdentity,
-    ) -> GraphDelta {
-        let runtime = SourceRuntimeId::parse(format!("{source}-prod")).unwrap();
-        let collection = CompleteCollection::new(
-            canonical.entity().tenant_id().clone(),
-            runtime.clone(),
-            CollectionId::parse(format!("{source}-{provider_id}")).unwrap(),
-            format!("{source}.users"),
-            20,
-        )
-        .unwrap();
-        let provider = ProviderIdentity::new(
-            canonical.entity().tenant_id().clone(),
-            runtime,
-            ProviderKind::parse(format!("{source}.identity_user")).unwrap(),
-            provider_id,
-            format!("{source} account"),
-        )
-        .unwrap();
-        let provenance = AssertionProvenance::direct(
-            vec![
-                ObservationRef::new(
-                    collection.receipt(),
-                    ObservationId::parse(format!("observation-{source}-{provider_id}")).unwrap(),
-                    format!("{source}.user:{provider_id}"),
-                )
-                .unwrap(),
-            ],
-            format!("{source}-identity-mapper"),
-            "v1",
-        )
-        .unwrap();
-        let binding = IdentityBindingAssertion::new(
-            &provider,
-            canonical,
-            IdentityResolutionMethod::ExistingClaimMatch,
-            Some(IdentityClaim::verified_email(email).unwrap()),
-            IdentityBindingState::Confirmed,
-            provenance,
-            20,
-        )
-        .unwrap();
-        let mut builder = collection.begin_delta();
-        builder.add_entity(provider.into_entity()).unwrap();
-        builder.add_entity(canonical.clone().into_entity()).unwrap();
-        builder
-            .add_assertion(GraphAssertion::IdentityBinding(binding))
-            .unwrap();
-        builder.build()
-    }
-
-    #[test]
-    fn entity_conflict_is_rejected_without_partial_mutation() {
-        let tenant = TenantId::parse("tenant-a").unwrap();
-        let runtime = SourceRuntimeId::parse("okta-prod").unwrap();
-        let provider = ProviderIdentity::new(
-            tenant.clone(),
-            runtime.clone(),
-            ProviderKind::parse("okta.user").unwrap(),
-            "00u-conflict",
-            "Provider Person",
-        )
-        .unwrap()
-        .into_entity();
-        let conflicting = Entity::canonical(
-            tenant.clone(),
-            provider.id().clone(),
-            EntityKind::Identity,
-            "Canonical collision",
-        )
-        .unwrap();
-
-        let mut graph = OrganizationalGraph::new();
-        let mut first = CompleteCollection::new(
-            tenant.clone(),
-            runtime.clone(),
-            CollectionId::parse("collection-conflict-first").unwrap(),
-            "okta.users",
-            10,
-        )
-        .unwrap()
-        .begin_delta();
-        first.add_entity(provider.clone()).unwrap();
-        graph.apply(first.build()).unwrap();
-        let revision = graph.graph_revision(&tenant);
-
-        let mut second = CompleteCollection::new(
-            tenant.clone(),
-            runtime,
-            CollectionId::parse("collection-conflict-second").unwrap(),
-            "okta.users",
-            10,
-        )
-        .unwrap()
-        .begin_delta();
-        second.add_entity(conflicting).unwrap();
-
-        assert_eq!(
-            graph.apply(second.build()),
-            Err(GraphError::EntityConflict(provider.id().clone()))
-        );
-        assert_eq!(graph.graph_revision(&tenant), revision);
-        assert_eq!(graph.entities(&tenant), vec![provider]);
-        assert!(graph.assertions(&tenant).is_empty());
-    }
-
-    #[test]
-    fn one_provider_identity_cannot_bind_to_two_canonical_identities() {
-        let mut graph = OrganizationalGraph::new();
-        let first = identity_delta("00u1", "person-1", IdentityBindingState::Confirmed, None);
-        let (provider_identity, existing_canonical_identity) = match &first.assertions()[0] {
-            GraphAssertion::IdentityBinding(binding) => (
-                binding.provider_identity().clone(),
-                binding.canonical_identity().clone(),
-            ),
-            _ => panic!("identity delta must contain an identity binding"),
-        };
-        graph.apply(first).unwrap();
-
-        let second = identity_delta("00u1", "person-2", IdentityBindingState::Confirmed, None);
-        let requested_canonical_identity = match &second.assertions()[0] {
-            GraphAssertion::IdentityBinding(binding) => binding.canonical_identity().clone(),
-            _ => panic!("identity delta must contain an identity binding"),
-        };
-        let revision = graph.graph_revision(&TenantId::parse("tenant-a").unwrap());
-
-        assert_eq!(
-            graph.apply(second),
-            Err(GraphError::IdentityAlreadyBound {
-                provider_identity,
-                existing_canonical_identity,
-                requested_canonical_identity,
-            })
-        );
-        let tenant = TenantId::parse("tenant-a").unwrap();
-        assert_eq!(graph.graph_revision(&tenant), revision);
-        assert_eq!(graph.assertions(&tenant).len(), 1);
-        assert_eq!(graph.entities(&tenant).len(), 2);
-    }
-
-    #[test]
-    fn stable_entity_identity_allows_source_data_refresh() {
-        let tenant = TenantId::parse("tenant-a").unwrap();
-        let runtime = SourceRuntimeId::parse("okta-prod").unwrap();
-        let entity = ProviderIdentity::new(
-            tenant.clone(),
-            runtime.clone(),
-            ProviderKind::parse("okta.user").unwrap(),
-            "00u-refresh",
-            "Old label",
-        )
-        .unwrap()
-        .into_entity();
-        let refreshed = ProviderIdentity::new(
-            tenant.clone(),
-            runtime.clone(),
-            ProviderKind::parse("okta.user").unwrap(),
-            "00u-refresh",
-            "New label",
-        )
-        .unwrap()
-        .into_entity()
-        .with_property("department", "Security")
-        .unwrap();
-        let mut graph = OrganizationalGraph::new();
-        for (collection_id, value) in [
-            ("collection-refresh-1", entity),
-            ("collection-refresh-2", refreshed.clone()),
-        ] {
-            let collection = CompleteCollection::new(
-                tenant.clone(),
-                runtime.clone(),
-                CollectionId::parse(collection_id).unwrap(),
-                "okta.users",
-                10,
-            )
-            .unwrap();
-            let mut builder = collection.begin_delta();
-            builder.add_entity(value).unwrap();
-            graph.apply(builder.build()).unwrap();
-        }
-        assert_eq!(graph.entity(&tenant, refreshed.id()).unwrap(), refreshed);
-    }
-
-    #[test]
-    fn one_authoritative_claim_cannot_create_two_canonical_identities() {
-        let mut graph = OrganizationalGraph::new();
-        let first = identity_delta(
-            "00u1",
-            "person-1",
-            IdentityBindingState::Confirmed,
-            Some(IdentityClaim::employee_id("employee-1").unwrap()),
-        );
-        let claim_kind = IdentityClaimKind::EmployeeId;
-        let claim_value = "employee-1".to_owned();
-        let existing_canonical_identity = match &first.assertions()[0] {
-            GraphAssertion::IdentityBinding(binding) => binding.canonical_identity().clone(),
-            _ => panic!("identity delta must contain an identity binding"),
-        };
-        let second = identity_delta(
-            "00u2",
-            "person-2",
-            IdentityBindingState::Confirmed,
-            Some(IdentityClaim::employee_id("employee-1").unwrap()),
-        );
-        let requested_canonical_identity = match &second.assertions()[0] {
-            GraphAssertion::IdentityBinding(binding) => binding.canonical_identity().clone(),
-            _ => panic!("identity delta must contain an identity binding"),
-        };
-        graph.apply(first).unwrap();
-        let revision = graph.graph_revision(&TenantId::parse("tenant-a").unwrap());
-        assert_eq!(
-            graph.apply(second),
-            Err(GraphError::IdentityClaimAlreadyBound {
-                claim_kind,
-                claim_value,
-                existing_canonical_identity,
-                requested_canonical_identity,
-            })
-        );
-        let tenant = TenantId::parse("tenant-a").unwrap();
-        assert_eq!(graph.graph_revision(&tenant), revision);
-        assert_eq!(graph.assertions(&tenant).len(), 1);
-        assert_eq!(graph.entities(&tenant).len(), 2);
-    }
-
-    #[test]
-    fn missing_assertion_entity_is_rejected_before_any_write() {
-        let tenant = TenantId::parse("tenant-a").unwrap();
-        let runtime = SourceRuntimeId::parse("okta-prod").unwrap();
-        let collection = CompleteCollection::new(
-            tenant.clone(),
-            runtime.clone(),
-            CollectionId::parse("collection-missing-entity").unwrap(),
-            "okta.users",
-            10,
-        )
-        .unwrap();
-        let provider = ProviderIdentity::new(
-            tenant.clone(),
-            runtime,
-            ProviderKind::parse("okta.user").unwrap(),
-            "00u-missing",
-            "Provider Person",
-        )
-        .unwrap()
-        .into_entity();
-        let repository = Entity::canonical(
-            tenant,
-            EntityId::parse("repository:missing").unwrap(),
-            EntityKind::Repository,
-            "Repository",
-        )
-        .unwrap();
-        let provenance = AssertionProvenance::direct(
-            vec![
-                ObservationRef::new(
-                    collection.receipt(),
-                    ObservationId::parse("observation-missing-entity").unwrap(),
-                    "okta.user:00u-missing",
-                )
-                .unwrap(),
-            ],
-            "okta-identity-mapper",
-            "v1",
-        )
-        .unwrap();
-        let relationship = RelationshipAssertion::new(
-            &provider,
-            RelationKind::CanAccess,
-            &repository,
-            provenance,
-            10,
-        )
-        .unwrap();
-        let mut builder = collection.begin_delta();
-        builder
-            .add_assertion(GraphAssertion::Relationship(relationship))
-            .unwrap();
-
-        let mut graph = OrganizationalGraph::new();
-        assert_eq!(
-            graph.apply(builder.build()),
-            Err(GraphError::MissingEntity(provider.id().clone()))
-        );
-        assert_eq!(
-            graph.graph_revision(&TenantId::parse("tenant-a").unwrap()),
-            0
-        );
-        assert!(
-            graph
-                .entities(&TenantId::parse("tenant-a").unwrap())
-                .is_empty()
-        );
-        assert!(
-            graph
-                .assertions(&TenantId::parse("tenant-a").unwrap())
-                .is_empty()
-        );
-    }
-
-    #[test]
-    fn verified_email_binding_requires_an_authoritative_anchor() {
-        let mut graph = OrganizationalGraph::new();
-        let delta = identity_delta_with_method(
-            "00u-unanchored",
-            "person-unanchored",
-            IdentityResolutionMethod::VerifiedEmail,
-            IdentityBindingState::Confirmed,
-            Some(IdentityClaim::verified_email("person@example.com").unwrap()),
-        );
-        let canonical_identity = match &delta.assertions()[0] {
-            GraphAssertion::IdentityBinding(binding) => binding.canonical_identity().clone(),
-            _ => panic!("identity delta must contain an identity binding"),
-        };
-
-        assert_eq!(
-            graph.apply(delta),
-            Err(GraphError::CanonicalIdentityUnanchored(canonical_identity))
-        );
-        let tenant = TenantId::parse("tenant-a").unwrap();
-        assert_eq!(graph.graph_revision(&tenant), 0);
-        assert!(graph.entities(&tenant).is_empty());
-        assert!(graph.assertions(&tenant).is_empty());
-    }
-
-    #[test]
-    fn existing_claim_match_requires_a_confirmed_claim() {
-        let tenant = TenantId::parse("tenant-a").unwrap();
-        let canonical = CanonicalIdentity::new(
-            tenant,
-            CanonicalIdentityId::parse("person-without-claim").unwrap(),
-            "Canonical Person",
-        )
-        .unwrap();
-        let delta = claim_match_delta(
-            "github",
-            "github-without-claim",
-            "missing@example.com",
-            &canonical,
-        );
-        let requested_canonical_identity = match &delta.assertions()[0] {
-            GraphAssertion::IdentityBinding(binding) => binding.canonical_identity().clone(),
-            _ => panic!("claim-match delta must contain an identity binding"),
-        };
-        let mut graph = OrganizationalGraph::new();
-
-        assert_eq!(
-            graph.apply(delta),
-            Err(GraphError::IdentityClaimNotFound {
-                claim_kind: IdentityClaimKind::VerifiedEmail,
-                claim_value: "missing@example.com".to_owned(),
-                requested_canonical_identity,
-            })
-        );
-        assert_eq!(
-            graph.graph_revision(&TenantId::parse("tenant-a").unwrap()),
-            0
-        );
-    }
-
-    #[test]
-    fn retraction_from_another_runtime_is_rejected_without_removal() {
-        let first = identity_delta(
-            "00u-retract",
-            "person-retract",
-            IdentityBindingState::Confirmed,
-            None,
-        );
-        let assertion_id = first.assertions()[0].id().clone();
-        let tenant = TenantId::parse("tenant-a").unwrap();
-        let mut graph = OrganizationalGraph::new();
-        graph.apply(first).unwrap();
-        let revision = graph.graph_revision(&tenant);
-
-        let collection = CompleteCollection::new(
-            tenant.clone(),
-            SourceRuntimeId::parse("github-prod").unwrap(),
-            CollectionId::parse("github-retract").unwrap(),
-            "github.users",
-            10,
-        )
-        .unwrap();
-        let mut builder = collection.begin_delta();
-        builder
-            .retract_missing(assertion_id.clone(), "provider no longer reports it")
-            .unwrap();
-
-        assert_eq!(
-            graph.apply(builder.build()),
-            Err(GraphError::RetractionSourceMismatch(assertion_id))
-        );
-        assert_eq!(graph.graph_revision(&tenant), revision);
-        assert_eq!(graph.assertions(&tenant).len(), 1);
-    }
-
-    #[test]
-    fn failed_delta_does_not_partially_change_graph() {
-        let mut graph = OrganizationalGraph::new();
-        graph
-            .apply(identity_delta(
-                "00u1",
-                "person-1",
-                IdentityBindingState::Confirmed,
-                None,
-            ))
-            .unwrap();
-        let revision = graph.graph_revision(&TenantId::parse("tenant-a").unwrap());
-        let _ = graph.apply(identity_delta(
-            "00u1",
-            "person-2",
-            IdentityBindingState::Confirmed,
-            None,
-        ));
-        assert_eq!(
-            graph.graph_revision(&TenantId::parse("tenant-a").unwrap()),
-            revision
-        );
-        assert_eq!(
-            graph.entities(&TenantId::parse("tenant-a").unwrap()).len(),
-            2,
-            "a rejected delta cannot insert its new canonical identity"
-        );
-    }
-
-    #[test]
-    fn workforce_claims_unify_okta_github_and_slack_accounts() {
-        let mut graph = OrganizationalGraph::new();
-        let (workforce, canonical) = workforce_delta("employee-1", "person@example.com");
-        graph.apply(workforce).unwrap();
-        graph
-            .apply(claim_match_delta(
-                "github",
-                "github-1",
-                "person@example.com",
-                &canonical,
-            ))
-            .unwrap();
-        graph
-            .apply(claim_match_delta(
-                "slack",
-                "slack-1",
-                "person@example.com",
-                &canonical,
-            ))
-            .unwrap();
-        let tenant = TenantId::parse("tenant-a").unwrap();
-        assert_eq!(
-            graph
-                .entities(&tenant)
-                .iter()
-                .filter(|entity| entity.kind() == &cerebro_organizational_model::EntityKind::Person)
-                .count(),
-            1
-        );
-    }
-
-    #[test]
-    fn shared_alias_cannot_create_or_redirect_a_person() {
-        let mut graph = OrganizationalGraph::new();
-        let (workforce, canonical) = workforce_delta("employee-1", "person@example.com");
-        graph.apply(workforce).unwrap();
-        assert!(matches!(
-            graph.apply(claim_match_delta(
-                "slack",
-                "slack-2",
-                "shared@example.com",
-                &canonical,
-            )),
-            Err(GraphError::IdentityClaimNotFound { .. })
-        ));
-    }
-
-    #[test]
-    fn non_directory_provider_cannot_seed_verified_email() {
-        let tenant = TenantId::parse("tenant-a").unwrap();
-        let runtime = SourceRuntimeId::parse("slack-prod").unwrap();
-        let collection = CompleteCollection::new(
-            tenant.clone(),
-            runtime.clone(),
-            CollectionId::parse("slack-user-1").unwrap(),
-            "slack.users",
-            10,
-        )
-        .unwrap();
-        let provider = ProviderIdentity::new(
-            tenant.clone(),
-            runtime,
-            ProviderKind::parse("slack.identity_user").unwrap(),
-            "U1",
-            "Slack account",
-        )
-        .unwrap();
-        let employee_claim = IdentityClaim::employee_id("employee-1").unwrap();
-        let canonical =
-            CanonicalIdentity::for_claim(tenant, &employee_claim, "Canonical Person").unwrap();
-        let provenance = AssertionProvenance::direct(
-            vec![
-                ObservationRef::new(
-                    collection.receipt(),
-                    ObservationId::parse("observation-slack-1").unwrap(),
-                    "slack.user:U1",
-                )
-                .unwrap(),
-            ],
-            "slack-identity-mapper",
-            "v1",
-        )
-        .unwrap();
-        assert_eq!(
-            IdentityBindingAssertion::new(
-                &provider,
-                &canonical,
-                IdentityResolutionMethod::VerifiedEmail,
-                Some(IdentityClaim::verified_email("person@example.com").unwrap()),
-                IdentityBindingState::Confirmed,
-                provenance,
-                10,
-            ),
-            Err(cerebro_organizational_model::ModelError::InvalidIdentityBinding)
-        );
-    }
-}
+mod tests;
